refactor(jwt): reorganize packages and rename config class
