+++ core/security/jwt/blolet-jwt-filter/src/main/java/nettee/jwt/filter/JwtAuthorizationFilter.java
@@ -11,7 +11,7 @@ import java.time.Instant;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import nettee.jwt.JwtParser;
+import nettee.jwt.parser.JwtParser;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
