+++ core/security/jwt/jwt-issuer/src/main/java/nettee/jwt/issuer/JwtIssuer.java
@@ -1,4 +1,4 @@
-package nettee.jwt;
+package nettee.jwt.issuer;
 
 import io.jsonwebtoken.JwtBuilder;
 import io.jsonwebtoken.Jwts;
