+++ core/security/jwt/jwt-issuer/src/main/java/nettee/jwt/issuer/JwtIssuerConfig.java
@@ -1,12 +1,13 @@
-package nettee.jwt;
+package nettee.jwt.issuer;
 
+import nettee.jwt.JwtProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @EnableConfigurationProperties(JwtProperties.class)
-public class JwtConfig {
+public class JwtIssuerConfig {
 
     @Bean
     public JwtIssuer jwtIssuer(JwtProperties properties) {
