+++ core/security/jwt/jwt-parser/src/main/java/nettee/jwt/parser/JwtParser.java
@@ -1,4 +1,4 @@
-package nettee.jwt;
+package nettee.jwt.parser;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtException;
