+++ core/security/jwt/jwt-parser/src/main/java/nettee/jwt/parser/JwtParserConfig.java
@@ -1,12 +1,13 @@
-package nettee.jwt;
+package nettee.jwt.parser;
 
+import nettee.jwt.JwtProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @EnableConfigurationProperties(JwtProperties.class)
-public class JwtConfig {
+public class JwtParserConfig {
 
     @Bean
     public JwtParser jwtParser(JwtProperties properties) {
