+++ core/security/jwt/jwt-parser/src/test/kotlin/nettee/jwt/JwtParserTest.kt
@@ -4,6 +4,8 @@ import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.assertions.throwables.shouldNotThrowAny
 import io.jsonwebtoken.Claims
+import nettee.jwt.issuer.JwtIssuer
+import nettee.jwt.parser.JwtParser
 import java.security.KeyPairGenerator
 import java.util.*
